+++ CHANGELOG.rst
@@ -148,6 +148,8 @@ Removed
 
 Fixed
 -----
+- Default ``EBSD.detector.shape`` is now correct when a detector is not passed upon
+  initialization. (`#603 <https://github.com/pyxem/kikuchipy/pull/603>`_)
 - Oxford Instruments .ebsp files of version 4 can now be read.
   (`#602 <https://github.com/pyxem/kikuchipy/pull/602>`_)
 - When loading EBSD patterns from H5OINA files, the detector tilt and binning are